Backend/requirements.txt
Flask
torch
Pillow 